♻️ Extract Shipping endpoint host into ShippingEndpointHost
